refactor: FinancialCalendar 컴포넌트 추가 및 로직 리팩토링
